Restructure role to an own Section.
