+++ config/roles.yaml
@@ -1,0 +1,28 @@
+services:
+  _defaults:
+    autowire: true
+    autoconfigure: true
+    public: false
+
+  # controllers are imported separately to make sure they're public
+  # and have a tag that allows actions to type-hint services
+  Pimcore\Bundle\StudioBackendBundle\Role\Controller\:
+    resource: '../src/Role/Controller'
+    public: true
+    tags: [ 'controller.service_arguments' ]
+
+  #
+  # Services
+  #
+  Pimcore\Bundle\StudioBackendBundle\Role\Service\RoleServiceInterface:
+    class: Pimcore\Bundle\StudioBackendBundle\Role\Service\RoleService
+
+  #
+  # Hydrators
+  #
+
+  #
+  # Repositories
+  #
+  Pimcore\Bundle\StudioBackendBundle\Role\Repository\RoleRepositoryInterface:
+    class: Pimcore\Bundle\StudioBackendBundle\Role\Repository\RoleRepository
