+++ config/users.yaml
@@ -35,9 +35,6 @@ services:
   Pimcore\Bundle\StudioBackendBundle\User\Service\WorkspaceCloneServiceInterface:
     class: Pimcore\Bundle\StudioBackendBundle\User\Service\WorkspaceCloneService
 
-  Pimcore\Bundle\StudioBackendBundle\User\Service\RoleServiceInterface:
-    class: Pimcore\Bundle\StudioBackendBundle\User\Service\RoleService
-
   Pimcore\Bundle\StudioBackendBundle\User\Service\MailServiceInterface:
     class: Pimcore\Bundle\StudioBackendBundle\User\Service\MailService
     arguments: [ '@Pimcore\Bundle\StudioBackendBundle\Setting\Provider\SystemSettingsProvider' ]
@@ -65,6 +62,3 @@ services:
 
   Pimcore\Bundle\StudioBackendBundle\User\Repository\PermissionRepositoryInterface:
     class: Pimcore\Bundle\StudioBackendBundle\User\Repository\PermissionRepository
-
-  Pimcore\Bundle\StudioBackendBundle\User\Repository\RoleRepositoryInterface:
-    class: Pimcore\Bundle\StudioBackendBundle\User\Repository\RoleRepository
