+++ src/Role/Schema/UserRole.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\StudioBackendBundle\User\Schema;
+namespace Pimcore\Bundle\StudioBackendBundle\Role\Schema;
 
 use OpenApi\Attributes\Property;
 use OpenApi\Attributes\Schema;
