+++ src/Role/Service/RoleServiceInterface.php
@@ -14,20 +14,18 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\StudioBackendBundle\User\Repository;
+namespace Pimcore\Bundle\StudioBackendBundle\Role\Service;
 
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\DatabaseException;
-use Pimcore\Model\User\Role;
+use Pimcore\Bundle\StudioBackendBundle\Response\Collection;
 
 /**
  * @internal
  */
-interface RoleRepositoryInterface
+interface RoleServiceInterface
 {
     /**
-     * @return Role[]
-     *
      * @throws DatabaseException
      */
-    public function getRoles(): array;
+    public function getRoles(): Collection;
 }
